@@ -1,25 +1,6 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { AppContainer } from 'react-hot-loader'
-import App from './components/App'
-
-ReactDOM.render(
-  <AppContainer>
-    <App/>
-  </AppContainer>,
-  document.getElementById('root')
-);
-
-// Hot Module Replacement API
-if (module.hot) {
-  module.hot.accept('./components/App', () => {
-    const NextApp = require('./components/App').default;
-    ReactDOM.render(
-      <AppContainer>
-        <NextApp/>
-      </AppContainer>
-      ,
-      document.getElementById('root')
-    );
-  });
-}
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './components/App';
+ 
+const rootElement = document.getElementById('root');
+ReactDOM.render(<App />, rootElement);
