@@ -1,7 +1,7 @@
 var webpack = require('webpack');
 
 module.exports = {
-    entry: ['react-hot-loader/patch', './src/index.js'] ,
+    entry: './src/index.js',
 
     output: {
         path: __dirname + '/public/',
@@ -16,16 +16,15 @@ module.exports = {
         contentBase: __dirname + '/public/',
     },
 
-    module:{
+    module: {
         loaders: [
             {
-                test: /.js$/,
+                test: /\.js$/,
                 loader: 'babel',
                 exclude: /node_modules/,
                 query: {
                     cacheDirectory: true,
-                    presets: ['es2015', 'react'],
-                    plugins: ["react-hot-loader/babel"]
+                    presets: ['es2015', 'react']
                 }
             }
         ]
@@ -34,4 +33,4 @@ module.exports = {
     plugins: [
         new webpack.HotModuleReplacementPlugin()
     ]
-};
+}
